refactor: 투표 응답 DTO를 기능별로 list/와 result/ 하위 패키지로 분리 (#36)
