+++ src/main/java/com/moa/moa_server/domain/vote/dto/response/list/VoteListResponse.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.domain.vote.dto.response;
+package com.moa.moa_server.domain.vote.dto.response.list;
 
 import java.util.List;
