+++ src/main/java/com/moa/moa_server/domain/vote/dto/response/result/VoteOptionResult.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.domain.vote.dto.response;
+package com.moa.moa_server.domain.vote.dto.response.result;
 
 public record VoteOptionResult(
         int optionNumber,
