+++ src/main/java/com/moa/moa_server/domain/vote/dto/response/result/VoteResultResponse.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.domain.vote.dto.response;
+package com.moa.moa_server.domain.vote.dto.response.result;
 
 import java.util.List;
