+++ src/main/java/com/moa/moa_server/domain/vote/service/VoteService.java
@@ -12,8 +12,8 @@ import com.moa.moa_server.domain.user.util.AuthUserValidator;
 import com.moa.moa_server.domain.vote.dto.request.VoteCreateRequest;
 import com.moa.moa_server.domain.vote.dto.request.VoteSubmitRequest;
 import com.moa.moa_server.domain.vote.dto.response.VoteDetailResponse;
-import com.moa.moa_server.domain.vote.dto.response.VoteOptionResult;
-import com.moa.moa_server.domain.vote.dto.response.VoteResultResponse;
+import com.moa.moa_server.domain.vote.dto.response.result.VoteOptionResult;
+import com.moa.moa_server.domain.vote.dto.response.result.VoteResultResponse;
 import com.moa.moa_server.domain.vote.entity.Vote;
 import com.moa.moa_server.domain.vote.entity.VoteResponse;
 import com.moa.moa_server.domain.vote.handler.VoteErrorCode;
